restructure kernel directory
