+++ src/squlearn/kernel/loss/negative_log_likelihood.py
@@ -5,7 +5,7 @@ import numpy as np
 
 from typing import Sequence
 from .kernel_loss_base import KernelLossBase
-from ..matrix.kernel_matrix_base import KernelMatrixBase
+from ..lowlevel_kernel.kernel_matrix_base import KernelMatrixBase
 
 
 class NLL(KernelLossBase):
