+++ src/squlearn/kernel/lowlevel_kernel/fidelity_kernel.py
@@ -267,7 +267,7 @@ class FidelityKernel(KernelMatrixBase):
         kernel_matrix = self._quantum_kernel.evaluate(x, y)
 
         if self._mit_depol_noise is not None:
-            print("WARNING: Advanced option. Do not use it within an squlearn.kernel.ml workflow")
+            print("WARNING: Advanced option. Do not use it within an squlearn.kernel workflow")
             if not np.array_equal(x, y):
                 raise ValueError(
                     "Mitigating depolarizing noise works only for square matrices computed on real"
