+++ src/squlearn/kernel/lowlevel_kernel/kernel_optimizer.py
