+++ src/squlearn/kernel/lowlevel_kernel/kernel_util.py
