+++ src/squlearn/kernel/lowlevel_kernel/regularization.py
