--- src/squlearn/kernel/ml/__init__.py
@@ -1,8 +1,0 @@
-from .qgpr import QGPR
-from .qgpc import QGPC
-from .qkrr import QKRR
-from .qsvr import QSVR
-from .qsvc import QSVC
-
-
-__all__ = ["QGPR", "QKRR", "QGPC", "QSVR", "QSVC"]
